docs: popover docs fixes
