+++ package.json
@@ -11,6 +11,7 @@
   "private": true,
   "scripts": {
     "serve": "yarn workspace vuestic-ui serve",
+    "serve:book": "yarn workspace vuestic-ui serve",
     "test": "yarn workspace vuestic-ui test",
     "lint": "yarn workspace vuestic-ui lint",
     "build": "yarn workspace vuestic-ui build",
