+++ packages/docs/src/components/page-configs/ui-elements/va-popover/page-config.ts
@@ -1,6 +1,6 @@
-import { BlockType, ApiDocsBlock } from '../../../types/configTypes'
 import VaPopover from 'vuestic-ui/src/components/vuestic-components/va-popover/VaPopover.vue'
 import apiOptions from './api-options'
+import { BlockType, ApiDocsBlock } from '@/types/configTypes'
 
 export default [
   {
