+++ packages/docs/src/examples/va-popover/Icon.vue
